+++ bin/act_config.ml
@@ -1,0 +1,14 @@
+(* The Automagic Compiler Tormentor
+
+   Copyright (c) 2018--2019 Matt Windsor and contributors
+
+   ACT itself is licensed under the MIT License. See the LICENSE file in the
+   project root for more information.
+
+   ACT is based in part on code from the Herdtools7 project
+   (https://github.com/herd/herdtools7) : see the LICENSE.herd file in the
+   project root for more information. *)
+
+(** Main entry point for the ACT config inspection services. *)
+
+let () = Core.Command.run Cmd_config.Main.command